task 19

+++ Task19/Program.cs
@@ -9,7 +9,7 @@ namespace Task19
             Console.WriteLine("Enter the number end it's power please: ");
             string[] inputs = Console.ReadLine().Split();
 
-            double n = double.Parse(inputs[0]);
+            int n = int.Parse(inputs[0]);
             int pow = int.Parse(inputs[1]);
  
             Console.WriteLine(Math.Pow(n, pow));
